+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.0 - 2024-10-10]
+## [1.1.0 - 2024-10-30]
 ### Added
 - Java Opts
 - Divide Excel file in several files if number of rows is too large
@@ -60,6 +60,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Java 21
 - Zip filename with SITE and TIMESTAMP
 - Refactoring Dockerfile
+- OPAL CSV (replacement of quotes with escape charachters)
 
 
 ### Removed